+++ jamanchu/module-batch/src/test/java/com/recipe/jamanchu/batch/BatchApplicationTests.java
@@ -1,0 +1,13 @@
+package com.recipe.jamanchu.batch;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BatchApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
