+++ jamanchu/module-domain/src/test/java/com/recipe/jamanchu/domain/DomainApplicationTests.java
@@ -1,0 +1,13 @@
+package com.recipe.jamanchu.domain;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DomainApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
