--- jamanchu/src/test/java/com/recipe/jamanchu/JamanchuApplicationTests.java
@@ -1,13 +1,0 @@
-package com.recipe.jamanchu;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JamanchuApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
